requirements.txt
asgiref==3.7.2
Django==4.2.2
django-filter==23.2
django-jazzmin==2.6.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
Markdown==3.4.3
phonenumbers==8.13.14
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
